Revert "Port nailgun client to rust (#10865)"
This reverts commit ee50791d77ee69305a55e9d5b47fed12076411f4.

[ci skip-build-wheels]
