+++ src/rust/engine/src/scheduler.rs
@@ -572,7 +572,7 @@ impl Scheduler {
   }
 }
 
-pub fn maybe_break_execution_loop(python_signal_fn: &Value) -> Option<ExecutionTermination> {
+fn maybe_break_execution_loop(python_signal_fn: &Value) -> Option<ExecutionTermination> {
   match externs::call_function(&python_signal_fn, &[]) {
     Ok(value) => {
       if externs::is_truthy(&value) {
